Extract job model and fix meta files
