+++ Assets/testgenetics/ScheduleCrossover.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 3a00524f-6ba5-40d3-8c75-db4f1beccdff
+guid: 488827ba9fdf42aa8fe1857c26fa2594
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
-  userData:
-  assetBundleName:
-  assetBundleVariant:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
